+++ .github/workflows/events-processor-tests.yml
@@ -10,7 +10,7 @@ on:
       - synchronize
       - reopened
     paths:
-      - "events_processor/**"
+      - "events_processors/**"
 
 jobs:
   test:
@@ -18,7 +18,7 @@ jobs:
     runs-on: ubuntu-latest
     defaults:
       run:
-        working-directory: ./events_processor
+        working-directory: ./events_processors
 
     services:
       postgres:
