+++ events_processors/config/kafka/consumer.go
@@ -9,7 +9,7 @@ import (
 	"github.com/twmb/franz-go/pkg/kgo"
 	"go.opentelemetry.io/otel/attribute"
 
-	tracer "github.com/getlago/lago/events-processor/config"
+	tracer "github.com/getlago/lago/events-processors/config"
 )
 
 type ConsumerGroupConfig struct {
