+++ events_processors/database/billable_metrics.go
@@ -5,7 +5,7 @@ import (
 
 	"gorm.io/gorm"
 
-	"github.com/getlago/lago/events-processor/utils"
+	"github.com/getlago/lago/events-processors/utils"
 )
 
 type BillableMetric struct {
