+++ events_processors/utils/result.go
